changed precision of the EvaluatorFunction interface

+++ Java/src/WatchmakerExtension/src/se/miun/itm/watchmaker/EvaluatorFunction.java
@@ -22,7 +22,7 @@ public interface EvaluatorFunction<T> {
 
 	boolean isNatural();
 	
-	float evaluate(T candidate);
+	double evaluate(T candidate);
 
 	int getRuns();
 }
